+++ libs/nav/include/mrpt/nav/tpspace/CPTG_DiffDrive_CollisionGridBased.h
@@ -69,7 +69,7 @@ namespace nav
 		size_t getPathStepCount(uint16_t k) const MRPT_OVERRIDE;
 		void getPathPose(uint16_t k, uint16_t step, mrpt::math::TPose2D &p) const MRPT_OVERRIDE;
 		double getPathDist(uint16_t k, uint16_t step) const MRPT_OVERRIDE;
-		bool getPathStepForDist(uint16_t k, double dist, uint16_t &out_step) const MRPT_OVERRIDE;
+		bool getPathStepForDist(uint16_t k, double dist, uint32_t &out_step) const MRPT_OVERRIDE;
 		double getPathStepDuration() const MRPT_OVERRIDE;
 		double getMaxLinVel() const MRPT_OVERRIDE { return V_MAX; }
 		double getMaxAngVel() const MRPT_OVERRIDE { return W_MAX; }
